Extract optimistic indexing code to separate module for better reusability
